refactor: useStartPageOptions
