+++ paths.json
@@ -14,6 +14,8 @@
       "@organismsNew": ["src/components/organisms_new/"],
       "@molecules/*": ["src/components/molecules/*"],
       "@molecules": ["src/components/molecules/"],
+      "@moleculesNew/*": ["src/components/molecules_new/*"],
+      "@moleculesNew": ["src/components/molecules_new/"],
       "@hooks/*": ["src/hooks/*"],
       "@hooks": ["src/hooks"],
       "@redux/*": ["src/redux/*"],
