+++ src/components/molecules_new/index.ts
@@ -1,1 +1,2 @@
 export {default as ProjectCard} from './ProjectCard';
+export {default as ProjectsList} from './ProjectsList';
