+++ src/components/organisms_new/SettingsPane/SettingsPane.styled.tsx
@@ -11,10 +11,6 @@ export const SettingsPaneContainer = styled.div<{$isOnStartProjectPage: boolean}
   background-color: ${({$isOnStartProjectPage}) => ($isOnStartProjectPage ? 'transparent' : '#191f21')};
 `;
 
-export const DescriptionContainer = styled.div`
-  padding: 8px 0px;
-`;
-
 export const WalkThroughContainer = styled.div`
   display: flex;
   align-items: center;
@@ -46,49 +42,6 @@ export const CurlyArrowImage = styled.img`
   border-radius: 100%;
   padding: 16px;
   margin-right: 10px;
-`;
-
-export const LayoutOption = styled.div<{$selected?: boolean}>`
-  border-radius: 4px;
-  display: flex;
-  justify-content: space-between;
-  height: 70px;
-  cursor: pointer;
-  background: ${({$selected}) => ($selected ? 'rgba(255, 255, 255, 0.2)' : 'inherit')};
-  border: ${({$selected}) => ($selected ? `2px solid ${Colors.blue7}` : `1px solid ${Colors.blue11}`)};
-  padding: ${({$selected}) => ($selected ? '3px' : '4px')};
-`;
-
-export const LayoutContainer = styled.div`
-  margin-right: 12px;
-  padding: 4px 0 4px 4px;
-
-  @media (max-width: 900px) {
-    display: flex;
-    align-items: center;
-  }
-`;
-
-export const LayoutTitle = styled.div`
-  color: ${Colors.whitePure};
-  font-size: 12px;
-  font-weight: 700;
-`;
-
-export const OptionsContainer = styled.div`
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  grid-gap: 10px;
-`;
-
-export const LayoutContent = styled.div`
-  color: ${Colors.whitePure};
-  font-size: 12px;
-  display: none;
-
-  @media (min-width: 925px) {
-    display: block;
-  }
 `;
 
 export const ThemeOption = styled.div<{$selected?: boolean}>`
