+++ src/components/organisms_new/SettingsPane/TitleCardDescription/index.ts
@@ -1,0 +1,1 @@
+export {default} from './TitleCardDescription';
