+++ src/components/organisms_new/SettingsPane/index.ts
@@ -1,1 +1,1 @@
-export {SettingsPane} from './SettingsPane';
+export {default} from './SettingsPane';
