+++ src/components/organisms_new/index.ts
@@ -3,5 +3,6 @@ export {default as LearnPage} from './LearnPage';
 export {default as NavigatorPane} from './NavigatorPane';
 export {default as NewProject} from './NewProject';
 export {default as PaneManager} from './PaneManager';
+export {default as SettingsPane} from './SettingsPane';
 export {default as StartPage} from './StartPage';
 export {default as ValidationSettings} from './ValidationSettings';
